feat(style): replace lodash with fp-ts everywhere
fix(api): require aws creds
