+++ package.json
@@ -4,7 +4,9 @@
   "description": "Define a single source of truth for your API and persistence layer, and use Verdad's extensions to a) call it, b) implement it, and c) deploy it.",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
-  "files": ["lib"],
+  "files": [
+    "lib"
+  ],
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc",
@@ -22,14 +24,13 @@
     "@tsconfig/node16-strictest": "^1.0.1",
     "@types/aws-lambda": "^8.10.97",
     "@types/console-log-level": "^1.4.2",
-    "@types/lodash": "^4.14.182",
     "@types/luxon": "1.25.0",
     "@types/node": "^17.0.31",
     "@types/safe-json-stringify": "^1.1.2",
     "@types/uuid": "^8.3.4",
     "serverless-plugin-typescript": "^2.1.2",
-    "typescript": "^4.6.4",
-    "standard-version": "^9.5.0"
+    "standard-version": "^9.5.0",
+    "typescript": "^4.6.4"
   },
   "dependencies": {
     "@aws-sdk/client-dynamodb": "^3.85.0",
@@ -38,10 +39,10 @@
     "axios": "^0.27.2",
     "console-log-level": "^1.4.1",
     "fp-ts": "^2.12.1",
+    "fp-ts-std": "^0.14.2",
     "http-status-codes": "^2.2.0",
     "io-ts": "^2.2.16",
     "io-ts-types": "^0.5.16",
-    "lodash": "^4.17.21",
     "luxon": "^1.28.0",
     "monocle-ts": "^2.3.13",
     "newtype-ts": "^0.3.5",
